Added `SDK.REST` tests for null at select and expand parameters Also moved fakeRequests.ts out of XrmQuery folder to have it easily used by other tests as well
