+++ test/src/tests/utils/init.ts
@@ -10,8 +10,8 @@ global.GetGlobalContext = () => ({ getClientUrl: () => "" });
 // Setup JavaScript dependencies
 const libPath = __dirname + "/../../../lib/";
 
-execFile(libPath + "sdk.rest.js");
 execFile(libPath + "sdk.metadata.js");
+execFile(libPath + "sdk.rest.js");
 execFile(libPath + "dg.xrmquery.web.js");
 execFile(libPath + "dg.xrmquery.rest.js");
